@@ -1,5 +1,4 @@
-class Stock < ActiveRecord::Base
-
+class Stock < ApplicationRecord
   has_many :user_stocks
   has_many :users, through: :user_stocks
 
@@ -10,6 +9,8 @@ class Stock < ActiveRecord::Base
   def self.new_from_lookup(ticker_symbol)
     begin
       looked_up_stock = StockQuote::Stock.quote(ticker_symbol)
+
+
       price = strip_commas(looked_up_stock.l)
       new(name: looked_up_stock.name, ticker: looked_up_stock.symbol, last_price: price)
     rescue Exception => e
@@ -20,4 +21,5 @@ class Stock < ActiveRecord::Base
   def self.strip_commas(number)
     number.gsub(",", "")
   end
+
 end
