@@ -1,4 +1,4 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
@@ -13,6 +13,7 @@ class User < ActiveRecord::Base
     "Anonymous"
   end
 
+
   def stock_already_added?(ticker_symbol)
     stock = Stock.find_by_ticker(ticker_symbol)
     return false unless stock
@@ -25,6 +26,14 @@ class User < ActiveRecord::Base
 
   def can_add_stock?(ticker_symbol)
     under_stock_limit? && !stock_already_added?(ticker_symbol)
+  end
+
+  def except_current_user(users)
+    users.reject { |user| user.id == self.id }
+  end
+
+  def not_friends_with?(friend_id)
+    friendships.where(friend_id: friend_id).count < 1
   end
 
   def self.search(param)
@@ -49,13 +58,5 @@ class User < ActiveRecord::Base
 
   def self.matches(field_name, param)
     where("#{field_name} like ?", "%#{param}%")
-  end
-
-  def except_current_user(users)
-    users.reject { |user| user.id == self.id}
-  end
-
-  def not_friends_with?(friend_id)
-    friendships.where(friend_id: friend_id).count < 1
   end
 end
