@@ -1,4 +1,4 @@
-class UserStock < ActiveRecord::Base
+class UserStock < ApplicationRecord
   belongs_to :user
   belongs_to :stock
 end
