@@ -11,7 +11,7 @@
 // about supported directives.
 //
 //= require jquery
-//= require jquery_ujs
+//= require rails-ujs
 //= require twitter/bootstrap
 // require turbolinks
 //= require_tree .
